refactor(core): organize utils
