+++ apps/web/src/app/(main)/_components/currently-reading-novels.tsx
@@ -1,8 +1,8 @@
 import { CardContainer } from '@/components/card-container'
 import { NovelCard } from '@/components/novel-card'
 import { getLatestReadNovels } from '@/lib/actions/novels'
-import { slugify } from '@/lib/utils'
 
+import { slugify } from '@yomu/core/utils/string'
 import { Progress } from '@yomu/ui/components/progress'
 
 import Link from 'next/link'
