+++ apps/web/src/app/(main)/novels/[id]/[slug]/_components/chapter-item.tsx
@@ -1,5 +1,5 @@
 import { Chapter } from '@yomu/core/database/schema/web'
-import { toCalendar } from '@yomu/core/date-helpers'
+import { toCalendar } from '@yomu/core/utils/dates'
 import { Input } from '@yomu/ui/components/input'
 import { Label } from '@yomu/ui/components/label'
 import { cn } from '@yomu/ui/utils'
