+++ packages/core/src/utils/string.ts
